split to book.rs
